Sincronizando produtores e consumidor

+++ main.c
@@ -19,7 +19,7 @@ struct shared_area
 {
     sem_t mutex1; // Para os produtores escreverem
     sem_t mutex2; // Para regular retirada da fila entre t1 e t2 de P4
-    pid_t paused_process;
+    int ready_for_produce;
     int queue_size;
     int queue[QUEUESIZE];
 };
@@ -113,11 +113,6 @@ int main()
 
         shared_area_ptr_fifo1 = (struct shared_area *)shared_memory_fifo1;
 
-        if (ready_for_pickup != 1)
-        {
-            pause();
-        }
-
         pthread_create(&threads[0], NULL, handle_t1, (void *)shared_area_ptr_fifo1);
         pthread_create(&threads[1], NULL, handle_t2, (void *)shared_area_ptr_fifo1);
 
@@ -140,7 +135,6 @@ int main()
         }
         if (produtores[i] == 0)
         {
-            signal(SIGUSR2, signal_handler_produtores);
             srand(time(NULL) + getpid());
             shmid_fifo1 = shmget(fifo1, MEM_SZ, 0666 | IPC_CREAT);
 
@@ -162,33 +156,35 @@ int main()
 
             if (i == 0)
             {
-                shared_area_ptr_fifo1->queue_size = 0;
                 if (sem_init((sem_t *)&shared_area_ptr_fifo1->mutex1, 1, 1) != 0)
-                { // Cria um semáforo pra ser utilizado entre processos
+                { // Semáforo dos processos
                     printf("sem_init falhou (FIFO 1, SEMAFORO 1)\n");
                     exit(-1);
                 }
                 if (sem_init((sem_t *)&shared_area_ptr_fifo1->mutex2, 0, 1) != 0)
-                {
+                { // Semáforo das threads
                     printf("sem_init falhou (FIFO 1, SEMAFORO 2)\n");
                     exit(-1);
                 }
+                sem_wait((sem_t *)&shared_area_ptr_fifo1->mutex1);
+                shared_area_ptr_fifo1->queue_size = 0;
+                shared_area_ptr_fifo1->ready_for_produce = 1;
+                sem_post((sem_t *)&shared_area_ptr_fifo1->mutex1);
             }
 
             for (;;)
             {
+                if (process4 == 0) continue;
                 sem_wait((sem_t *)&shared_area_ptr_fifo1->mutex1);
-                if (shared_area_ptr_fifo1->queue_size < 10)
+                if (shared_area_ptr_fifo1->queue_size < 10 && shared_area_ptr_fifo1->ready_for_produce == 1)
                 {
                     int x = rand_interval(MINNUMBER, MAXNUMBER);
-                    shared_area_ptr_fifo1->queue[shared_area_ptr_fifo1->queue_size] = x; // Insere na última posição da fila
-                    shared_area_ptr_fifo1->queue_size += 1;                              // Soma um no tamanho atual
-
+                    shared_area_ptr_fifo1->queue[shared_area_ptr_fifo1->queue_size] = x;
+                    shared_area_ptr_fifo1->queue_size += 1;
                     if (shared_area_ptr_fifo1->queue_size == 10)
                     {
-                        shared_area_ptr_fifo1->paused_process = getpid();
+                        shared_area_ptr_fifo1->ready_for_produce = 0;
                         kill(process4, SIGUSR1);
-                        pause();
                     }
                 }
                 sem_post((sem_t *)&shared_area_ptr_fifo1->mutex1);
@@ -308,7 +304,7 @@ int main()
     if (process7 == -1)
     {
         printf("Erro ao criar P7\n");
-        exit(-1); 
+        exit(-1);
     }
     else if (process7 == 0)
     {
@@ -392,9 +388,9 @@ void *handle_t1(void *ptr)
     for (;;)
     {
         sem_wait(&shared_area_ptr_fifo1->mutex2);
-
         if (ready_for_pickup == 1)
         {
+            sem_wait(&shared_area_ptr_fifo1->mutex1);
             if (shared_area_ptr_fifo1->queue_size > 0)
             {
                 int res = shared_area_ptr_fifo1->queue[0]; // Pega o primeiro da fila
@@ -410,9 +406,10 @@ void *handle_t1(void *ptr)
                 if (shared_area_ptr_fifo1->queue_size == 0)
                 {
                     ready_for_pickup = 0;
-                    kill(shared_area_ptr_fifo1->paused_process, SIGUSR2);
+                    shared_area_ptr_fifo1->ready_for_produce = 1;
                 }
             }
+            sem_post(&shared_area_ptr_fifo1->mutex1);
         }
         sem_post(&shared_area_ptr_fifo1->mutex2);
     }
@@ -426,9 +423,9 @@ void *handle_t2(void *ptr)
     for (;;)
     {
         sem_wait(&shared_area_ptr_fifo1->mutex2);
-
         if (ready_for_pickup == 1)
         {
+            sem_wait(&shared_area_ptr_fifo1->mutex1);
             if (shared_area_ptr_fifo1->queue_size > 0)
             {
                 int res = shared_area_ptr_fifo1->queue[0]; // Pega o primeiro da fila
@@ -444,9 +441,10 @@ void *handle_t2(void *ptr)
                 if (shared_area_ptr_fifo1->queue_size == 0)
                 {
                     ready_for_pickup = 0;
-                    kill(shared_area_ptr_fifo1->paused_process, SIGUSR2);
+                    shared_area_ptr_fifo1->ready_for_produce = 1;
                 }
             }
+            sem_post(&shared_area_ptr_fifo1->mutex1);
         }
         sem_post(&shared_area_ptr_fifo1->mutex2);
     }
@@ -527,11 +525,6 @@ void *handle_threads_p7(void *ptr)
 void signal_handler_consumidores(int p)
 {
     ready_for_pickup = 1;
-}
-
-void signal_handler_produtores(int p)
-{
-    raise(SIGCONT);
 }
 
 int rand_interval(int a, int b)
